+++ packages/spacecat-shared-content-client/src/clients/content-client.js
@@ -264,8 +264,8 @@ export default class ContentClient {
     this.log.info(`Getting page metadata for ${this.site.getId()} and path ${path}`);
 
     const docPath = this.#resolveDocPath(path);
-    const document = await this.rawClient.read(docPath);
-    const metadata = document.getMetadata();
+    const document = this.rawClient.getDocument(docPath);
+    const metadata = await document.getMetadata();
 
     this.#logDuration('getPageMetadata', startTime);
 
@@ -284,8 +284,8 @@ export default class ContentClient {
     this.log.info(`Updating page metadata for ${this.site.getId()} and path ${path}`);
 
     const docPath = this.#resolveDocPath(path);
-    const document = await this.rawClient.read(docPath);
-    const originalMetadata = document.getMetadata();
+    const document = this.rawClient.getDocument(docPath);
+    const originalMetadata = await document.getMetadata();
 
     let mergedMetadata;
     if (overwrite) {
@@ -296,10 +296,7 @@ export default class ContentClient {
 
     const response = await document.updateMetadata(mergedMetadata);
     if (response?.status !== 200) {
-      const saveResponse = await this.rawClient.save(document);
-      if (saveResponse.status !== 200) {
-        throw new Error(`Failed to update metadata for path ${path}`);
-      }
+      throw new Error(`Failed to update metadata for path ${path}`);
     }
 
     this.#logDuration('updatePageMetadata', startTime);
@@ -313,7 +310,8 @@ export default class ContentClient {
 
     this.log.info(`Getting redirects for ${this.site.getId()}`);
 
-    const redirects = await this.rawClient.getRedirects();
+    const redirectsFile = this.rawClient.getRedirects();
+    const redirects = await redirectsFile.get();
     this.#logDuration('getRedirects', startTime);
 
     return redirects;
@@ -328,8 +326,8 @@ export default class ContentClient {
 
     this.log.info(`Updating redirects for ${this.site.getId()}`);
 
-    const currentRedirects = await this.getRedirects();
-
+    const redirectsFile = this.rawClient.getRedirects();
+    const currentRedirects = await redirectsFile.get();
     // validate combination of existing and new redirects
     const cleanNewRedirects = removeDuplicatedRedirects(currentRedirects, redirects, this.log);
     if (cleanNewRedirects.length === 0) {
@@ -342,7 +340,7 @@ export default class ContentClient {
       return;
     }
 
-    const response = await this.rawClient.appendRedirects(noCycleRedirects);
+    const response = await redirectsFile.append(noCycleRedirects);
     if (response.status !== 200) {
       throw new Error('Failed to update redirects');
     }